backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-storages==1.14.2
django-environ==0.11.2
psycopg==3.1.18
Pillow==11.2.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
boto3==1.34.0
python-decouple==3.8
django-ratelimit==4.1.0
django-cleanup==8.1.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-silk==5.1.0
django-cacheops==7.2
django-redis==5.4.0
django-ses==3.5.0
django-imagekit==5.0.0
django-taggit==4.0.0
whitenoise==6.6.0
gunicorn==21.2.0
requests==2.31.0
python-magic==0.4.27
ffmpeg-python==0.2.0
django-allauth==0.61.1
dj-rest-auth==5.0.1
djangorestframework-simplejwt==5.3.1
django-guardian==2.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1 